Exception handler

+++ src/main/java/com/bogdan/vocabulary/service/dictionary/DictionaryServiceImpl.java
@@ -1,6 +1,8 @@
 package com.bogdan.vocabulary.service.dictionary;
 
 import com.bogdan.vocabulary.dto.DictionaryDto;
+import com.bogdan.vocabulary.exception.dictionary.DictionaryValidationException;
+import com.bogdan.vocabulary.exception.dictionary.DictionaryNotFoundException;
 import com.bogdan.vocabulary.mapper.DictionaryMapper;
 import com.bogdan.vocabulary.model.Dictionary;
 import com.bogdan.vocabulary.repository.DictionaryRepository;
@@ -12,6 +14,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
 @Service
@@ -27,6 +30,13 @@ public class DictionaryServiceImpl implements DictionaryService {
 
     @Override
     public DictionaryDto createDictionary(DictionaryDto dictionaryDto) {
+        boolean isValid = Pattern.matches(
+                "^[a-zA-Z0-9]+('[a-zA-Z0-9])?[a-zA-Z0-9]{0,36}", dictionaryDto.getDictionaryName());
+
+        if (!isValid) {
+            throw new DictionaryValidationException("Invalid 'name' format.");
+        }
+
         Dictionary dictionary = DictionaryMapper.mapToDictionary(dictionaryDto);
         Dictionary savedDictionary = dictionaryRepository.save(dictionary);
         return DictionaryMapper.mapToDictionaryDto(savedDictionary);
@@ -44,11 +54,17 @@ public class DictionaryServiceImpl implements DictionaryService {
     public DictionaryDto patchDictionary(Long id, Map<String, Object> changes) {
         Optional<Dictionary> optionalDictionary = dictionaryRepository.findById(id);
 
-        Dictionary savedDictionary = optionalDictionary.orElse(null);
+        if (optionalDictionary.isEmpty()) {
+            throw new DictionaryNotFoundException("Dictionary with id: " + id + " not found.");
+        }
 
-        if (savedDictionary == null) {
-            return null;
-            //todo
+        Dictionary savedDictionary = optionalDictionary.get();
+
+        boolean isValid = Pattern.matches(
+                "^[a-zA-Z0-9]+('[a-zA-Z0-9])?[a-zA-Z0-9]{0,36}", savedDictionary.getDictionaryName());
+
+        if (!isValid) {
+            throw new DictionaryValidationException("Invalid 'name' format.");
         }
 
         changes.forEach((change, value) -> {
@@ -64,6 +80,12 @@ public class DictionaryServiceImpl implements DictionaryService {
 
     @Override
     public void deleteDictionary(Long id) {
+        Optional<Dictionary> optionalDictionary = dictionaryRepository.findById(id);
+
+        if (optionalDictionary.isEmpty()) {
+            throw new DictionaryNotFoundException("Dictionary with id: " + id + " not found.");
+        }
+
         dictionaryRepository.deleteById(id);
     }
 }
